requirements.txt
streamlit>=1.44
pandas>=2.2
pydantic>=2.7
PyMuPDF>=1.24
PyYAML>=6.0
python-dotenv>=1.0
sentence-transformers>=5.4
